fix(useCase):updated use case and model for save total amount

+++ backEnd/src/main/java/com/fasterEnvios/application/dto/packages/PackageRequestDTO.java
@@ -3,6 +3,8 @@ package com.fasterEnvios.application.dto.packages;
 import jakarta.validation.constraints.Negative;
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
+
 public record PackageRequestDTO (
         @NotBlank(message = "debes colocar el peso de este paquete")
         @Negative(message = "el peso no puede ser negativo")
@@ -12,7 +14,7 @@ public record PackageRequestDTO (
         double dimensions,
         @NotBlank(message = "debes colocar el valor declarado de este paquete")
         @Negative(message = "el valor no pueden ser negativo")
-        float declaredValue,
+        BigDecimal declaredValue,
         @NotBlank(message = "debes poner una descripcion valida")
         String description
 ){
